+++ src/rovr/core/file_list.py
@@ -3,18 +3,19 @@ import contextlib
 from os import getcwd, path
 from os import system as cmd
 from time import time
-from typing import ClassVar
+from typing import ClassVar, Iterable, Self
 
 from rich.segment import Segment
 from rich.style import Style
 from textual import events, on, work
 from textual.binding import BindingType
+from textual.content import ContentText
 from textual.css.query import NoMatches
 from textual.geometry import Region
 from textual.strip import Strip
 from textual.widgets import Button, Input, OptionList, SelectionList
 from textual.widgets.option_list import Option, OptionDoesNotExist
-from textual.widgets.selection_list import Selection
+from textual.widgets.selection_list import Selection, SelectionType
 
 from rovr.classes import ArchiveFileListSelection, FileListSelectionWidget
 from rovr.classes.session_manager import SessionManager
@@ -23,7 +24,12 @@ from rovr.functions import icons as icon_utils
 from rovr.functions import path as path_utils
 from rovr.functions import pins as pin_utils
 from rovr.functions import utils
-from rovr.variables.constants import buttons_that_depend_on_path, config, vindings
+from rovr.variables.constants import (
+    SortByOptions,
+    buttons_that_depend_on_path,
+    config,
+    vindings,
+)
 
 
 class FileList(SelectionList, inherit_bindings=False):
@@ -60,14 +66,19 @@ class FileList(SelectionList, inherit_bindings=False):
             self.input: Input = self.parent.query_one(Input)
 
     @property
-    def sort_by(self) -> str:
+    def sort_by(
+        self,
+    ) -> SortByOptions:
         try:
             return self.app.query_one("StateManager").sort_by
         except (NoMatches, AttributeError):
             return "name"
 
     @sort_by.setter
-    def sort_by(self, value: str) -> None:
+    def sort_by(
+        self,
+        value: SortByOptions,
+    ) -> None:
         if value not in ["name", "size", "modified", "created", "extension", "natural"]:
             raise ValueError(
                 f"Expected sort_by value to be one of 'name', 'size', 'modified', 'created', 'extension' or 'natural', but got '{value}'"
@@ -152,9 +163,13 @@ class FileList(SelectionList, inherit_bindings=False):
         cwd = path_utils.normalise(getcwd())
         # get sessionstate
         try:
-            # only happens when the tabs aren't mounted
             session: SessionManager = self.app.tabWidget.active_tab.session
         except AttributeError:
+            # only happens when the tabs aren't mounted
+            # this means that some stupid thing happened
+            # and i dont want filelist to die as well
+            # because it will be called later on (because of
+            # the watcher function)
             self.clear_options()
             return
         self.app.file_list_pause_check = True
@@ -173,7 +188,7 @@ class FileList(SelectionList, inherit_bindings=False):
                 folders, files = await path_utils.get_cwd_object(
                     cwd,
                     config["settings"]["show_hidden_files"],
-                    sort_by=self.sort_by,  # ty: ignore[invalid-argument-type]
+                    sort_by=self.sort_by,
                     reverse=self.sort_descending,
                 )
                 if not folders and not files:
@@ -225,12 +240,18 @@ class FileList(SelectionList, inherit_bindings=False):
             )
             for button in buttons:
                 button.disabled = should_disable
-            self.app.query_one("#new").disabled = self.list_of_options[0].id == "perm"
+            if len(self.list_of_options) > 0:
+                self.app.query_one("#new").disabled = (
+                    self.list_of_options[0].id == "perm"
+                )
+            else:
+                # this shouldnt happen, but just in case
+                self.app.query_one("#new").disabled = True
+
             # special check for up tree
             self.app.query_one("#up").disabled = cwd == path.dirname(cwd)
 
-            self.clear_options()
-            self.add_options(self.list_of_options)
+            self.set_options(self.list_of_options)
             # session handler
             self.app.query_one("#path_switcher").value = cwd + (
                 "" if cwd.endswith("/") else "/"
@@ -350,8 +371,7 @@ class FileList(SelectionList, inherit_bindings=False):
                     disabled=True,
                 )
             )
-        self.clear_options()
-        self.add_options(self.list_of_options)
+        self.set_options(self.list_of_options)
         self.parent.border_subtitle = ""
 
     @work(exclusive=True)
@@ -362,7 +382,6 @@ class FileList(SelectionList, inherit_bindings=False):
             file_list (list[str]): List of file paths from archive contents.
         """
         assert self.parent is not None
-        self.clear_options()
         self.list_of_options = []
 
         if not file_list:
@@ -391,7 +410,7 @@ class FileList(SelectionList, inherit_bindings=False):
                     start_time = time()
                 await asyncio.sleep(0)
 
-        self.add_options(self.list_of_options)
+        self.set_options(self.list_of_options)
         self.parent.border_subtitle = ""
 
     async def file_selected_handler(self, target_path: str) -> None:
@@ -667,9 +686,55 @@ class FileList(SelectionList, inherit_bindings=False):
 
     async def on_key(self, event: events.Key) -> None:
         """Handle key events for the file list."""
+        if self.dummy:
+            return
         from rovr.functions.utils import check_key
 
-        if not self.dummy and self.highlighted_option:
+        # hit buttons with keybinds
+        if not self.select_mode_enabled and check_key(
+            event, config["keybinds"]["hist_previous"]
+        ):
+            if self.app.query_one("#back").disabled:
+                self.app.query_one("UpButton").on_button_pressed(Button.Pressed)
+            else:
+                self.app.query_one("BackButton").on_button_pressed(Button.Pressed)
+        elif (
+            not self.select_mode_enabled
+            and check_key(event, config["keybinds"]["hist_next"])
+            and not self.app.query_one("#forward").disabled
+        ):
+            self.app.query_one("ForwardButton").on_button_pressed(Button.Pressed)
+        elif not self.select_mode_enabled and check_key(
+            event, config["keybinds"]["up_tree"]
+        ):
+            self.app.query_one("UpButton").on_button_pressed(Button.Pressed)
+        # Toggle pin on current directory
+        elif check_key(event, config["keybinds"]["toggle_pin"]):
+            pin_utils.toggle_pin(path.basename(getcwd()), getcwd())
+            await self.app.query_one("PinnedSidebar").reload_pins()
+        elif check_key(event, config["keybinds"]["copy"]):
+            await self.app.query_one("#copy").on_button_pressed(Button.Pressed)
+        elif check_key(event, config["keybinds"]["cut"]):
+            await self.app.query_one("#cut").on_button_pressed(Button.Pressed)
+        elif check_key(event, config["keybinds"]["paste"]):
+            await self.app.query_one("#paste").on_button_pressed(Button.Pressed)
+        elif check_key(event, config["keybinds"]["new"]):
+            self.app.query_one("#new").on_button_pressed(Button.Pressed)
+        elif check_key(event, config["keybinds"]["rename"]):
+            self.app.query_one("#rename").on_button_pressed(Button.Pressed)
+        elif check_key(event, config["keybinds"]["delete"]):
+            await self.app.query_one("#delete").on_button_pressed(Button.Pressed)
+        elif check_key(event, config["keybinds"]["zip"]):
+            self.app.query_one("#zip").on_button_pressed(Button.Pressed)
+        elif check_key(event, config["keybinds"]["unzip"]):
+            self.app.query_one("#unzip").on_button_pressed(Button.Pressed)
+        # search
+        elif check_key(event, config["keybinds"]["focus_search"]):
+            self.input.focus()
+        # toggle hidden files
+        elif check_key(event, config["keybinds"]["toggle_hidden_files"]):
+            await self.toggle_hidden_files()
+        elif self.highlighted_option:
             # toggle select mode
             if check_key(event, config["keybinds"]["toggle_visual"]):
                 await self.toggle_mode()
@@ -790,54 +855,10 @@ class FileList(SelectionList, inherit_bindings=False):
                             cmd,
                             f'{config["plugins"]["editor"]["file_executable"]} "{self.highlighted_option.dir_entry.path}"',
                         )
-            # hit buttons with keybinds
-            elif not self.select_mode_enabled and check_key(
-                event, config["keybinds"]["hist_previous"]
-            ):
-                if self.app.query_one("#back").disabled:
-                    self.app.query_one("UpButton").on_button_pressed(Button.Pressed)
-                else:
-                    self.app.query_one("BackButton").on_button_pressed(Button.Pressed)
-            elif (
-                not self.select_mode_enabled
-                and check_key(event, config["keybinds"]["hist_next"])
-                and not self.app.query_one("#forward").disabled
-            ):
-                self.app.query_one("ForwardButton").on_button_pressed(Button.Pressed)
-            elif not self.select_mode_enabled and check_key(
-                event, config["keybinds"]["up_tree"]
-            ):
-                self.app.query_one("UpButton").on_button_pressed(Button.Pressed)
             elif check_key(event, config["keybinds"]["copy_path"]):
                 await self.app.query_one("PathCopyButton").on_button_pressed(
                     Button.Pressed
                 )
-            # Toggle pin on current directory
-            elif check_key(event, config["keybinds"]["toggle_pin"]):
-                pin_utils.toggle_pin(path.basename(getcwd()), getcwd())
-                await self.app.query_one("PinnedSidebar").reload_pins()
-            elif check_key(event, config["keybinds"]["copy"]):
-                await self.app.query_one("#copy").on_button_pressed(Button.Pressed)
-            elif check_key(event, config["keybinds"]["cut"]):
-                await self.app.query_one("#cut").on_button_pressed(Button.Pressed)
-            elif check_key(event, config["keybinds"]["paste"]):
-                await self.app.query_one("#paste").on_button_pressed(Button.Pressed)
-            elif check_key(event, config["keybinds"]["new"]):
-                self.app.query_one("#new").on_button_pressed(Button.Pressed)
-            elif check_key(event, config["keybinds"]["rename"]):
-                self.app.query_one("#rename").on_button_pressed(Button.Pressed)
-            elif check_key(event, config["keybinds"]["delete"]):
-                await self.app.query_one("#delete").on_button_pressed(Button.Pressed)
-            elif check_key(event, config["keybinds"]["zip"]):
-                self.app.query_one("#zip").on_button_pressed(Button.Pressed)
-            elif check_key(event, config["keybinds"]["unzip"]):
-                self.app.query_one("#unzip").on_button_pressed(Button.Pressed)
-            # search
-            elif check_key(event, config["keybinds"]["focus_search"]):
-                self.input.focus()
-            # toggle hidden files
-            elif check_key(event, config["keybinds"]["toggle_hidden_files"]):
-                await self.toggle_hidden_files()
 
     def update_border_subtitle(self) -> None:
         if self.dummy or type(self.highlighted) is not int or not self.parent:
@@ -905,6 +926,30 @@ class FileList(SelectionList, inherit_bindings=False):
                 top=top,
                 immediate=True,
             )
+
+    def set_options(
+        self,
+        options: Iterable[
+            Selection[SelectionType]
+            | tuple[ContentText, SelectionType]
+            | tuple[ContentText, SelectionType, bool]
+        ],
+    ) -> Self:  # ty: ignore[invalid-method-override]
+        # Okay, lemme make myself clear here.
+        # A PR for this is already open at
+        # https://github.com/Textualize/textual/pull/6224
+        # essentially, the issue is that there isnt a set_options
+        # method for SelectionList, only for OptionList, but using
+        # OptionList's set_options doesnt clear selected or values
+        # but nothing was done, so I added it myself.
+        self._selected.clear()
+        self._values.clear()
+        # the ty ignore is important here, because options
+        # should be a Iterable["Option | VisualType | None"]
+        # but that isnt the case (based on the signature)
+        # so ty is crashing out.
+        super().set_options(options)  # ty: ignore[invalid-argument-type]
+        return self
 
 
 class FileListRightClickOptionList(PopupOptionList):