+++ CHANGELOG.md
@@ -1,3 +1,27 @@
+# [v0.6.0rc1]
+
+## new features
+- clipboard: constantly check clipboard added files 81df523
+- config: allow changing bindings for screen layers #161
+- fd!: rename from 'finder' to 'fd' #163
+- fd: add additional toggleable options #163
+- icons: show icon for symlink/junctions e6a354a
+- sort_order!: add custom keybind support #168
+- icons: show separate symlink/junction icon fbf2a08
+
+## build
+- pip: switch to tomli for toml parsing #162
+
+## fixes
+- filelist: fix issue with empty directories preventing navigation 985a509
+- input: fix overscroll issue a8b5307
+
+## perf
+- filelist: use custom set_options method e6a354a
+
+## docs
+- screenshots: perhaps fix the broken fonts #166
+
 # [v0.6.0.dev2]
 
 ## breaking changes